backend_lead_hunter/requirements.txt
fastapi
uvicorn
httpx
openai
pydantic-settings
python-dotenv
aiogram
